softlock removal in specific areas
general rule is that if you can't save vaguely close to an item that can force you into a softlock, then you will not be forced to dive into it and pray. this currently includes:

- Warrior Shrine Sealed Tunnel
- Hydro Access Tunnel

these changes do NOT include:

- Thardus
- Control Tower
- Burn Dome (both items)
- Any item you can look at before committing to the softlock

as well, the Sunchamber Ghost Duel has been reverted to Easy Difficulty conditions, due to the Bomb Climb technique being outside the scope of the Normal Difficulty and the ability to do Flaahgra before getting Spider.
in the same vein, no more OoB required for Arbor Chamber, as I vastly misjudged how punishing a failure OoB can be.

experimental changes to a couple tough lava areas were added:

- Magma Pool scandash will never be required anymore. However, you will be expected to L-Jump from a tall underlava rock if you have Gravity and Space Jump
- Fiery Shores Morph Track now checks for Space Jump, Gravity, Grapple, or Morph&Bombs after reviewing a complaint about the scandash. Gravity (w/o SJ) and Grapple have not been tested entry methods.

as a final note, if Sunchamber (Ghosts) and Elite Research (Phazon) are misbehaving, my compacted logic failed and needs to be reverted. (I set them to just check if the item that triggers their event has been obtained + any extraneous requirements)